requirements.txt
python-dotenv
ipykernel
langchain
langgraph
pydantic
streamlit
langchain-core
langchain-groq
langchain-openai
langchain-community
atlassian-python-api